+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.1.20
+
+- Adds Wallet info component
+- Adds User info component
+- Adds User info expanded component
+
 ## v0.1.19
 
 - Add value prop to Select component